Handle split-dns in xml and make search last element in resolv-conf
In certain cases, the search domain is specified in a split-dns element
in the xml which should be handled in the same way as the regular dns
element is.

Furthermore, resolv.conf (see resolv.conf(5)) only handles the last
search domain entry therefore the read dns-suffix should be added to the
end of the resolv.conf. The most elegant solution would be to merge
existing search entries with one received from the VPN server, e.g.
`search vpn.local lan.local`, where lan.local was received from the
local dhcp.